Refactored Message extractors into their own package, updated test package names
