+++ src/main/MessageExtractor/MessageExtractorInterface.java
@@ -1,4 +1,4 @@
-package clientHandlers;
+package MessageExtractor;
 
 public interface MessageExtractorInterface{
     public PeripheralAction handleMessage(String input) throws UnexpectedActionException;
