+++ src/test/MessageExtractor/JSONHandlerTest.java
@@ -1,13 +1,10 @@
-package messageHandlers;
+package MessageExtractor;
 
-import clientHandlers.JSONExtractor;
-import clientHandlers.UnexpectedActionException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.junit.jupiter.api.Assertions.expectThrows;
 
 public class JSONHandlerTest{
     JSONExtractor jsonEx;
